+++ tests/interfaces/test_torch.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2021 Xanadu Quantum Technologies Inc.
+# Copyright 2018-2022 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ import pytest
 pytestmark = pytest.mark.torch
 
 torch = pytest.importorskip("torch")
+torch_functional = pytest.importorskip("torch.autograd.functional")
 torch_cuda = pytest.importorskip("torch.cuda")
 
 import pennylane as qml
@@ -25,7 +26,7 @@ from pennylane.gradients import finite_diff, param_shift
 from pennylane.interfaces import execute
 
 
-@pytest.mark.parametrize("interface", ["torch"])
+@pytest.mark.parametrize("interface", ["torch", "auto"])
 class TestTorchExecuteUnitTests:
     """Unit tests for torch execution"""
 
@@ -43,6 +44,7 @@ class TestTorchExecuteUnitTests:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute(
             [tape],
             dev,
@@ -56,9 +58,9 @@ class TestTorchExecuteUnitTests:
         for args in spy.call_args_list:
             assert args[1]["shift"] == [(np.pi / 4,)] * 2
 
-    def test_incorrect_mode(self, interface):
+    def test_incorrect_grad_on_execution(self, interface):
         """Test that an error is raised if a gradient transform
-        is used with mode=forward"""
+        is used with grad_on_execution=True"""
         a = torch.tensor([0.1, 0.2], requires_grad=True)
 
         dev = qml.device("default.qubit", wires=1)
@@ -69,13 +71,14 @@ class TestTorchExecuteUnitTests:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         with pytest.raises(
-            ValueError, match="Gradient transforms cannot be used with mode='forward'"
+            ValueError, match="Gradient transforms cannot be used with grad_on_execution=True"
         ):
             execute([tape], dev, gradient_fn=param_shift, mode="forward", interface=interface)[0]
 
-    def test_forward_mode_reuse_state(self, interface, mocker):
-        """Test that forward mode uses the `device.execute_and_gradients` pathway
+    def test_grad_on_execution_reuse_state(self, interface, mocker):
+        """Test that grad_on_execution uses the `device.execute_and_gradients` pathway
         while reusing the quantum state."""
         dev = qml.device("default.qubit", wires=1)
         spy = mocker.spy(dev, "execute_and_gradients")
@@ -88,6 +91,7 @@ class TestTorchExecuteUnitTests:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute(
             [tape],
             dev,
@@ -100,8 +104,8 @@ class TestTorchExecuteUnitTests:
         assert dev.num_executions == 1
         spy.assert_called()
 
-    def test_forward_mode(self, interface, mocker):
-        """Test that forward mode uses the `device.execute_and_gradients` pathway"""
+    def test_grad_on_execution(self, interface, mocker):
+        """Test that grad on execution uses the `device.execute_and_gradients` pathway"""
         dev = qml.device("default.qubit", wires=1)
         spy = mocker.spy(dev, "execute_and_gradients")
 
@@ -113,6 +117,7 @@ class TestTorchExecuteUnitTests:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute(
             [tape],
             dev,
@@ -125,8 +130,8 @@ class TestTorchExecuteUnitTests:
         assert dev.num_executions == 2
         spy.assert_called()
 
-    def test_backward_mode(self, interface, mocker):
-        """Test that backward mode uses the `device.batch_execute` and `device.gradients` pathway"""
+    def test_no_grad_on_execution(self, interface, mocker):
+        """Test that no grad on execution uses the `device.batch_execute` and `device.gradients` pathway"""
         dev = qml.device("default.qubit", wires=1)
         spy_execute = mocker.spy(qml.devices.DefaultQubit, "batch_execute")
         spy_gradients = mocker.spy(qml.devices.DefaultQubit, "gradients")
@@ -139,6 +144,7 @@ class TestTorchExecuteUnitTests:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute(
             [tape],
             dev,
@@ -171,9 +177,10 @@ class TestCaching:
                 qml.probs(wires=0)
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return execute(
                 [tape], dev, gradient_fn=param_shift, cachesize=cachesize, interface="torch"
-            )[0][0, 0]
+            )[0][0]
 
         params = torch.tensor([0.1, 0.2], requires_grad=True)
         res = cost(params, cachesize=2)
@@ -196,8 +203,9 @@ class TestCaching:
                 qml.probs(wires=0)
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return execute([tape], dev, gradient_fn=param_shift, cache=cache, interface="torch")[0][
-                0, 0
+                0
             ]
 
         custom_cache = {}
@@ -220,20 +228,21 @@ class TestCaching:
                 qml.probs(wires=0)
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return execute([tape], dev, gradient_fn=param_shift, cache=cache, interface="torch")[0][
-                0, 0
+                0
             ]
 
         # Without caching, 5 evaluations are required to compute
         # the Jacobian: 1 (forward pass) + (2 shifts * 2 params)
         params = torch.tensor([0.1, 0.2], requires_grad=True)
-        torch.autograd.functional.jacobian(lambda p: cost(p, cache=None), params)
+        torch_functional.jacobian(lambda p: cost(p, cache=None), params)
         assert dev.num_executions == 5
 
         # With caching, 5 evaluations are required to compute
         # the Jacobian: 1 (forward pass) + (2 shifts * 2 params)
         dev._num_executions = 0
-        torch.autograd.functional.jacobian(lambda p: cost(p, cache=True), params)
+        torch_functional.jacobian(lambda p: cost(p, cache=True), params)
         assert dev.num_executions == 5
 
     @pytest.mark.parametrize("num_params", [2, 3])
@@ -258,6 +267,7 @@ class TestCaching:
                 qml.var(qml.PauliZ(0) @ qml.PauliX(1))
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return execute(
                 [tape], dev, gradient_fn=param_shift, cache=cache, interface="torch", max_diff=2
             )[0]
@@ -294,9 +304,9 @@ class TestCaching:
         assert dev.num_executions == expected_runs_ideal
         assert expected_runs_ideal < expected_runs
 
-    def test_caching_adjoint_backward(self):
+    def test_caching_adjoint_no_grad_on_execution(self):
         """Test that caching reduces the number of adjoint evaluations
-        when mode=backward"""
+        when grad_on_execution=False"""
         dev = qml.device("default.qubit", wires=2)
         params = torch.tensor([0.1, 0.2, 0.3])
 
@@ -309,6 +319,7 @@ class TestCaching:
                 qml.expval(qml.PauliZ(1))
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return execute(
                 [tape],
                 dev,
@@ -322,13 +333,13 @@ class TestCaching:
         # Without caching, 3 evaluations are required.
         # 1 for the forward pass, and one per output dimension
         # on the backward pass.
-        torch.autograd.functional.jacobian(lambda x: cost(x, cache=None), params)
+        torch_functional.jacobian(lambda x: cost(x, cache=None), params)
         assert dev.num_executions == 3
 
         # With caching, only 2 evaluations are required. One
         # for the forward pass, and one for the backward pass.
         dev._num_executions = 0
-        torch.autograd.functional.jacobian(lambda x: cost(x, cache=True), params)
+        torch_functional.jacobian(lambda x: cost(x, cache=True), params)
         assert dev.num_executions == 2
 
 
@@ -358,6 +369,25 @@ execute_kwargs = [
         "gradient_kwargs": {"method": "adjoint_jacobian"},
         "interface": "torch",
     },
+    {"gradient_fn": param_shift, "interface": "auto"},
+    {
+        "gradient_fn": "device",
+        "mode": "forward",
+        "gradient_kwargs": {"method": "adjoint_jacobian", "use_device_state": False},
+        "interface": "auto",
+    },
+    {
+        "gradient_fn": "device",
+        "mode": "forward",
+        "gradient_kwargs": {"method": "adjoint_jacobian", "use_device_state": True},
+        "interface": "auto",
+    },
+    {
+        "gradient_fn": "device",
+        "mode": "backward",
+        "gradient_kwargs": {"method": "adjoint_jacobian"},
+        "interface": "auto",
+    },
 ]
 
 
@@ -380,17 +410,20 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliZ(0))
 
         tape1 = qml.tape.QuantumScript.from_queue(q1)
+
         with qml.queuing.AnnotatedQueue() as q2:
             qml.RY(a, wires=0)
             qml.RX(b, wires=0)
             qml.expval(qml.PauliZ(0))
 
         tape2 = qml.tape.QuantumScript.from_queue(q2)
+
         res = execute([tape1, tape2], dev, **execute_kwargs)
 
+        assert isinstance(res, list)
         assert len(res) == 2
-        assert res[0].shape == (1,)
-        assert res[1].shape == (1,)
+        assert res[0].shape == ()
+        assert res[1].shape == ()
 
     def test_scalar_jacobian(self, torch_device, execute_kwargs, tol):
         """Test scalar jacobian calculation by comparing two types of pipelines"""
@@ -402,6 +435,7 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
         res.backward()
 
@@ -412,6 +446,7 @@ class TestTorchExecuteIntegration:
                 qml.expval(qml.PauliZ(0))
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             dev = qml.device("default.qubit.autograd", wires=2)
             return dev.batch_execute([tape])[0]
 
@@ -437,18 +472,25 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliY(1))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
         assert tape.trainable_params == [1, 2]
 
-        assert isinstance(res, torch.Tensor)
-        assert res.shape == (2,)
+        assert isinstance(res, tuple)
+
+        assert isinstance(res[0], torch.Tensor)
+        assert res[0].shape == ()
+
+        assert isinstance(res[1], torch.Tensor)
+        assert res[1].shape == ()
 
         expected = torch.tensor(
             [np.cos(a_val), -np.cos(a_val) * np.sin(b_val)], device=torch_device
         )
-        assert torch.allclose(res.detach(), expected, atol=tol, rtol=0)
+        assert torch.allclose(res[0].detach(), expected[0], atol=tol, rtol=0)
+        assert torch.allclose(res[1].detach(), expected[1], atol=tol, rtol=0)
 
-        loss = torch.sum(res)
+        loss = res[0] + res[1]
 
         loss.backward()
         expected = torch.tensor(
@@ -471,17 +513,20 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliX(0))
 
         tape1 = qml.tape.QuantumScript.from_queue(q1)
+
         with qml.queuing.AnnotatedQueue() as q2:
             qml.RY(0.5, wires=0)
             qml.expval(qml.PauliZ(0))
 
         tape2 = qml.tape.QuantumScript.from_queue(q2)
+
         with qml.queuing.AnnotatedQueue() as q3:
             qml.RY(params[0], wires=0)
             qml.RX(params[1], wires=0)
             qml.expval(qml.PauliZ(0))
 
         tape3 = qml.tape.QuantumScript.from_queue(q3)
+
         res = sum(execute([tape1, tape2, tape3], dev, **execute_kwargs))
         expected = torch.tensor(1 + np.cos(0.5), dtype=res.dtype) + torch.cos(x) * torch.cos(y)
         expected = expected.to(device=res.device)
@@ -512,10 +557,11 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliY(1))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         assert tape.trainable_params == [0, 1]
 
         res = execute([tape], dev, **execute_kwargs)[0]
-        loss = torch.sum(res)
+        loss = res[0] + res[1]
         loss.backward()
 
         a_val = 0.54
@@ -529,11 +575,12 @@ class TestTorchExecuteIntegration:
         expected = torch.tensor(
             [np.cos(2 * a_val), -np.cos(2 * a_val) * np.sin(b_val)],
             device=torch_device,
-            dtype=res2.dtype,
+            dtype=res2[0].dtype,
         )
-        assert torch.allclose(res2.detach(), expected, atol=tol, rtol=0)
+        assert torch.allclose(res2[0].detach(), expected[0], atol=tol, rtol=0)
+        assert torch.allclose(res2[1].detach(), expected[1], atol=tol, rtol=0)
 
-        loss = torch.sum(res2)
+        loss = res2[0] + res2[1]
         loss.backward()
 
         expected = torch.tensor(
@@ -562,6 +609,7 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
 
         assert tape.trainable_params == [0, 2]
@@ -597,17 +645,30 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliZ(1))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
         assert tape.trainable_params == []
 
-        assert res.shape == (2,)
-        assert isinstance(res, torch.Tensor)
+        assert isinstance(res, tuple)
+        assert len(res) == 2
+
+        assert isinstance(res[0], torch.Tensor)
+        assert res[0].shape == ()
+
+        assert isinstance(res[1], torch.Tensor)
+        assert res[1].shape == ()
 
         with pytest.raises(
             RuntimeError,
             match="element 0 of tensors does not require grad and does not have a grad_fn",
         ):
-            res.backward()
+            res[0].backward()
+
+        with pytest.raises(
+            RuntimeError,
+            match="element 0 of tensors does not require grad and does not have a grad_fn",
+        ):
+            res[1].backward()
 
     @pytest.mark.parametrize(
         "U", [torch.tensor([[0.0, 1.0], [1.0, 0.0]]), np.array([[0.0, 1.0], [1.0, 0.0]])]
@@ -629,6 +690,7 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
         assert tape.trainable_params == [1]
 
@@ -645,14 +707,14 @@ class TestTorchExecuteIntegration:
 
         class U3(qml.U3):
             def expand(self):
+                tape = qml.tape.QuantumTape()
                 theta, phi, lam = self.data
                 wires = self.wires
-                return qml.tape.QuantumScript(
-                    [
-                        qml.Rot(lam, theta, -lam, wires=wires),
-                        qml.PhaseShift(phi + lam, wires=wires),
-                    ]
-                )
+                tape._ops += [
+                    qml.Rot(lam, theta, -lam, wires=wires),
+                    qml.PhaseShift(phi + lam, wires=wires),
+                ]
+                return tape
 
         dev = qml.device("default.qubit", wires=1)
         a = np.array(0.1)
@@ -665,6 +727,7 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliX(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
 
         expected = torch.tensor(
@@ -711,41 +774,60 @@ class TestTorchExecuteIntegration:
         x = torch.tensor(x_val, requires_grad=True, device=torch_device)
         y = torch.tensor(y_val, requires_grad=True, device=torch_device)
 
-        with qml.queuing.AnnotatedQueue() as q:
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            qml.probs(wires=[0])
-            qml.probs(wires=[1])
+        def circuit(x, y):
+            with qml.queuing.AnnotatedQueue() as q:
+                qml.RX(x, wires=[0])
+                qml.RY(y, wires=[1])
+                qml.CNOT(wires=[0, 1])
+                qml.probs(wires=[0])
+                qml.probs(wires=[1])
 
-        tape = qml.tape.QuantumScript.from_queue(q)
-        res = execute([tape], dev, **execute_kwargs)[0]
+            tape = qml.tape.QuantumScript.from_queue(q)
 
-        expected = torch.tensor(
-            [
-                [np.cos(x_val / 2) ** 2, np.sin(x_val / 2) ** 2],
-                [(1 + np.cos(x_val) * np.cos(y_val)) / 2, (1 - np.cos(x_val) * np.cos(y_val)) / 2],
-            ],
-            dtype=res.dtype,
+            return execute([tape], dev, **execute_kwargs)[0]
+
+        res = circuit(x, y)
+
+        expected_0 = torch.tensor(
+            [np.cos(x_val / 2) ** 2, np.sin(x_val / 2) ** 2],
+            dtype=res[0].dtype,
             device=torch_device,
         )
-        assert torch.allclose(res.detach(), expected, atol=tol, rtol=0)
 
-        loss = torch.sum(res)
-        loss.backward()
-        expected = torch.tensor(
+        expected_1 = torch.tensor(
             [
-                -np.sin(x_val) / 2
-                + np.sin(x_val) / 2
-                - np.sin(x_val) * np.cos(y_val) / 2
-                + np.cos(y_val) * np.sin(x_val) / 2,
-                -np.cos(x_val) * np.sin(y_val) / 2 + np.cos(x_val) * np.sin(y_val) / 2,
+                (1 + np.cos(x_val) * np.cos(y_val)) / 2,
+                (1 - np.cos(x_val) * np.cos(y_val)) / 2,
             ],
-            dtype=x.grad.dtype,
+            dtype=res[0].dtype,
             device=torch_device,
         )
-        assert torch.allclose(x.grad, expected[0], atol=tol, rtol=0)
-        assert torch.allclose(y.grad, expected[1], atol=tol, rtol=0)
+
+        assert torch.allclose(res[0], expected_0, atol=tol, rtol=0)
+        assert torch.allclose(res[1], expected_1, atol=tol, rtol=0)
+
+        jac = torch_functional.jacobian(circuit, (x, y))
+        dtype_jac = jac[0][0].dtype
+
+        res_0 = torch.tensor(
+            [-np.sin(x_val) / 2, np.sin(x_val) / 2], dtype=dtype_jac, device=torch_device
+        )
+        res_1 = torch.tensor([0.0, 0.0], dtype=dtype_jac, device=torch_device)
+        res_2 = torch.tensor(
+            [-np.sin(x_val) * np.cos(y_val) / 2, np.cos(y_val) * np.sin(x_val) / 2],
+            dtype=dtype_jac,
+            device=torch_device,
+        )
+        res_3 = torch.tensor(
+            [-np.cos(x_val) * np.sin(y_val) / 2, +np.cos(x_val) * np.sin(y_val) / 2],
+            dtype=dtype_jac,
+            device=torch_device,
+        )
+
+        assert torch.allclose(jac[0][0], res_0, atol=tol, rtol=0)
+        assert torch.allclose(jac[0][1], res_1, atol=tol, rtol=0)
+        assert torch.allclose(jac[1][0], res_2, atol=tol, rtol=0)
+        assert torch.allclose(jac[1][1], res_3, atol=tol, rtol=0)
 
     def test_ragged_differentiation(self, torch_device, execute_kwargs, tol):
         """Tests correct output shape and evaluation for a tape
@@ -759,46 +841,68 @@ class TestTorchExecuteIntegration:
         x = torch.tensor(x_val, requires_grad=True, device=torch_device)
         y = torch.tensor(y_val, requires_grad=True, device=torch_device)
 
-        with qml.queuing.AnnotatedQueue() as q:
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            qml.expval(qml.PauliZ(0))
-            qml.probs(wires=[1])
+        def circuit(x, y):
+            with qml.queuing.AnnotatedQueue() as q:
+                qml.RX(x, wires=[0])
+                qml.RY(y, wires=[1])
+                qml.CNOT(wires=[0, 1])
+                qml.expval(qml.PauliZ(0))
+                qml.probs(wires=[1])
 
-        tape = qml.tape.QuantumScript.from_queue(q)
-        res = execute([tape], dev, **execute_kwargs)[0]
+            tape = qml.tape.QuantumScript.from_queue(q)
 
-        expected = torch.tensor(
-            [
-                np.cos(x_val),
-                (1 + np.cos(x_val) * np.cos(y_val)) / 2,
-                (1 - np.cos(x_val) * np.cos(y_val)) / 2,
-            ],
-            dtype=res.dtype,
+            return execute([tape], dev, **execute_kwargs)[0]
+
+        res = circuit(x, y)
+
+        res_0 = torch.tensor(np.cos(x_val), dtype=res[0].dtype, device=torch_device)
+        res_1 = torch.tensor(
+            [(1 + np.cos(x_val) * np.cos(y_val)) / 2, (1 - np.cos(x_val) * np.cos(y_val)) / 2],
+            dtype=res[0].dtype,
             device=torch_device,
         )
-        assert torch.allclose(res.detach(), expected, atol=tol, rtol=0)
 
-        loss = torch.sum(res)
-        loss.backward()
-        expected = torch.tensor(
-            [
-                -np.sin(x_val)
-                + -np.sin(x_val) * np.cos(y_val) / 2
-                + np.cos(y_val) * np.sin(x_val) / 2,
-                -np.cos(x_val) * np.sin(y_val) / 2 + np.cos(x_val) * np.sin(y_val) / 2,
-            ],
-            dtype=x.grad.dtype,
+        assert isinstance(res, tuple)
+        assert len(res) == 2
+
+        assert torch.allclose(res[0], res_0, atol=tol, rtol=0)
+        assert torch.allclose(res[1], res_1, atol=tol, rtol=0)
+
+        jac = torch_functional.jacobian(circuit, (x, y))
+        dtype_jac = jac[0][0].dtype
+
+        res_0 = torch.tensor(
+            -np.sin(x_val),
+            dtype=dtype_jac,
             device=torch_device,
         )
-        assert torch.allclose(x.grad, expected[0], atol=tol, rtol=0)
-        assert torch.allclose(y.grad, expected[1], atol=tol, rtol=0)
+        res_1 = torch.tensor(
+            0.0,
+            dtype=dtype_jac,
+            device=torch_device,
+        )
+        res_2 = torch.tensor(
+            [-np.sin(x_val) * np.cos(y_val) / 2, np.cos(y_val) * np.sin(x_val) / 2],
+            dtype=dtype_jac,
+            device=torch_device,
+        )
+        res_3 = torch.tensor(
+            [-np.cos(x_val) * np.sin(y_val) / 2, +np.cos(x_val) * np.sin(y_val) / 2],
+            dtype=dtype_jac,
+            device=torch_device,
+        )
+
+        assert torch.allclose(jac[0][0], res_0, atol=tol, rtol=0)
+        assert torch.allclose(jac[0][1], res_1, atol=tol, rtol=0)
+        assert torch.allclose(jac[1][0], res_2, atol=tol, rtol=0)
+        assert torch.allclose(jac[1][1], res_3, atol=tol, rtol=0)
 
     def test_sampling(self, torch_device, execute_kwargs):
         """Test sampling works as expected"""
         if execute_kwargs["gradient_fn"] == "device" and execute_kwargs["mode"] == "forward":
             pytest.skip("Adjoint differentiation does not support samples")
+        if execute_kwargs["interface"] == "auto":
+            pytest.skip("Can't detect interface without a parametrized gate in the tape")
 
         dev = qml.device("default.qubit", wires=2, shots=10)
 
@@ -809,15 +913,24 @@ class TestTorchExecuteIntegration:
             qml.sample(qml.PauliX(1))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
 
-        assert res.shape == (2, 10)
-        assert isinstance(res, torch.Tensor)
+        assert isinstance(res, tuple)
+        assert len(res) == 2
+
+        assert isinstance(res[0], torch.Tensor)
+        assert res[0].shape == (10,)
+
+        assert isinstance(res[1], torch.Tensor)
+        assert res[1].shape == (10,)
 
     def test_sampling_expval(self, torch_device, execute_kwargs):
         """Test sampling works as expected if combined with expectation values"""
         if execute_kwargs["gradient_fn"] == "device" and execute_kwargs["mode"] == "forward":
             pytest.skip("Adjoint differentiation does not support samples")
+        if execute_kwargs["interface"] == "auto":
+            pytest.skip("Can't detect interface without a parametrized gate in the tape")
 
         dev = qml.device("default.qubit", wires=2, shots=10)
 
@@ -828,11 +941,13 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliX(1))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
 
         assert len(res) == 2
         assert isinstance(res, tuple)
         assert res[0].shape == (10,)
+        assert res[1].shape == ()
         assert isinstance(res[0], torch.Tensor)
         assert isinstance(res[1], torch.Tensor)
 
@@ -850,6 +965,7 @@ class TestTorchExecuteIntegration:
             qml.sample(qml.PauliZ(0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         res = execute([tape], dev, **execute_kwargs)[0]
 
         with pytest.raises(
@@ -871,6 +987,7 @@ class TestTorchExecuteIntegration:
             qml.expval(qml.PauliZ(wires=0))
 
         tape = qml.tape.QuantumScript.from_queue(q)
+
         tape = tape.expand()
         res1 = execute([tape], dev, **execute_kwargs)[0]
 
@@ -901,6 +1018,7 @@ class TestHigherOrderDerivatives:
                 qml.var(qml.PauliZ(0) @ qml.PauliX(1))
 
             tape1 = qml.tape.QuantumScript.from_queue(q1)
+
             with qml.queuing.AnnotatedQueue() as q2:
                 qml.RX(x[0], wires=0)
                 qml.RY(x[0], wires=1)
@@ -908,10 +1026,11 @@ class TestHigherOrderDerivatives:
                 qml.probs(wires=1)
 
             tape2 = qml.tape.QuantumScript.from_queue(q2)
+
             result = execute(
                 [tape1, tape2], dev, gradient_fn=param_shift, interface="torch", max_diff=2
             )
-            return result[0] + result[1][0, 0]
+            return result[0] + result[1][0]
 
         res = cost_fn(params)
         x, y = params.detach()
@@ -944,6 +1063,7 @@ class TestHigherOrderDerivatives:
                 qml.probs(wires=0)
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return torch.stack(
                 execute([tape], dev, gradient_fn=param_shift, interface="torch", max_diff=2)
             )
@@ -966,11 +1086,11 @@ class TestHigherOrderDerivatives:
         )
         assert torch.allclose(res.detach(), expected_res, atol=tol, rtol=0)
 
-        jac_fn = lambda x: torch.autograd.functional.jacobian(circuit, x, create_graph=True)
+        jac_fn = lambda x: torch_functional.jacobian(circuit, x, create_graph=True)
 
         g = jac_fn(x)
 
-        hess = torch.autograd.functional.jacobian(jac_fn, x)
+        hess = torch_functional.jacobian(jac_fn, x)
 
         expected_g = torch.tensor(
             [
@@ -1014,6 +1134,7 @@ class TestHigherOrderDerivatives:
                 qml.expval(qml.PauliZ(0))
 
             tape = qml.tape.QuantumScript.from_queue(q)
+
             return execute(
                 [tape],
                 dev,
@@ -1040,6 +1161,7 @@ class TestHigherOrderDerivatives:
                 qml.var(qml.PauliZ(0) @ qml.PauliX(1))
 
             tape1 = qml.tape.QuantumScript.from_queue(q1)
+
             with qml.queuing.AnnotatedQueue() as q2:
                 qml.RX(x[0], wires=0)
                 qml.RY(x[0], wires=1)
@@ -1047,10 +1169,11 @@ class TestHigherOrderDerivatives:
                 qml.probs(wires=1)
 
             tape2 = qml.tape.QuantumScript.from_queue(q2)
+
             result = execute(
                 [tape1, tape2], dev, gradient_fn=param_shift, max_diff=1, interface="torch"
             )
-            return result[0] + result[1][0, 0]
+            return result[0] + result[1][0]
 
         res = cost_fn(params)
         x, y = params.detach()
@@ -1100,7 +1223,8 @@ class TestHamiltonianWorkflows:
                 qml.expval(H2)
 
             tape = qml.tape.QuantumScript.from_queue(q)
-            return execute([tape], dev, **execute_kwargs)[0]
+
+            return torch.hstack(execute([tape], dev, **execute_kwargs)[0])
 
         return _cost_fn
 
@@ -1140,12 +1264,11 @@ class TestHamiltonianWorkflows:
 
         res = cost_fn(weights, coeffs1, coeffs2, dev=dev)
         expected = self.cost_fn_expected(weights, coeffs1, coeffs2)
-        assert np.allclose(res.detach(), expected, atol=tol, rtol=0)
+        assert np.allclose(res[0].detach(), expected[0], atol=tol, rtol=0)
+        assert np.allclose(res[1].detach(), expected[1], atol=tol, rtol=0)
 
         res = torch.hstack(
-            torch.autograd.functional.jacobian(
-                lambda *x: cost_fn(*x, dev=dev), (weights, coeffs1, coeffs2)
-            )
+            torch_functional.jacobian(lambda *x: cost_fn(*x, dev=dev), (weights, coeffs1, coeffs2))
         )
         expected = self.cost_fn_jacobian(weights, coeffs1, coeffs2)
         assert np.allclose(res.detach(), expected, atol=tol, rtol=0)
@@ -1158,12 +1281,11 @@ class TestHamiltonianWorkflows:
 
         res = cost_fn(weights, coeffs1, coeffs2, dev=dev)
         expected = self.cost_fn_expected(weights, coeffs1, coeffs2)
-        assert np.allclose(res.detach(), expected, atol=tol, rtol=0)
+        assert np.allclose(res[0].detach(), expected[0], atol=tol, rtol=0)
+        assert np.allclose(res[1].detach(), expected[1], atol=tol, rtol=0)
 
         res = torch.hstack(
-            torch.autograd.functional.jacobian(
-                lambda *x: cost_fn(*x, dev=dev), (weights, coeffs1, coeffs2)
-            )
+            torch_functional.jacobian(lambda *x: cost_fn(*x, dev=dev), (weights, coeffs1, coeffs2))
         )
         expected = self.cost_fn_jacobian(weights, coeffs1, coeffs2)
         assert np.allclose(res.detach(), expected, atol=tol, rtol=0)
